+++ src/object.ts
