+++ src/string.ts
